+++ toc.go
@@ -118,12 +118,7 @@ func (root *assetTree) writeGoMap(w io.Writer, nident int) (err error) {
 }
 
 func (root *assetTree) WriteAsGoMap(w io.Writer) (err error) {
-	_, err = fmt.Fprint(w, `type bintree struct {
-	Func     func() (*asset, error)
-	Children map[string]*bintree
-}
-
-var _bintree = `)
+	_, err = fmt.Fprint(w, tmplTypeBintree)
 	if err != nil {
 		return
 	}
@@ -132,42 +127,7 @@ var _bintree = `)
 }
 
 func writeTOCTree(w io.Writer, toc []Asset) error {
-	_, err := fmt.Fprintf(w, `// AssetDir returns the file names below a certain
-// directory embedded in the file by go-bindata.
-// For example if you run go-bindata on data/... and data contains the
-// following hierarchy:
-//     data/
-//       foo.txt
-//       img/
-//         a.png
-//         b.png
-// then AssetDir("data") would return []string{"foo.txt", "img"}
-// AssetDir("data/img") would return []string{"a.png", "b.png"}
-// AssetDir("foo.txt") and AssetDir("notexist") would return an error
-// AssetDir("") will return []string{"data"}.
-func AssetDir(name string) ([]string, error) {
-	node := _bintree
-	if len(name) != 0 {
-		cannonicalName := strings.Replace(name, "\\", "/", -1)
-		pathList := strings.Split(cannonicalName, "/")
-		for _, p := range pathList {
-			node = node.Children[p]
-			if node == nil {
-				return nil, &os.PathError{Op: "open", Path: name, Err: os.ErrNotExist}
-			}
-		}
-	}
-	if node.Func != nil {
-		return nil, &os.PathError{Op: "open", Path: name, Err: os.ErrNotExist}
-	}
-	rv := make([]string, 0, len(node.Children))
-	for childName := range node.Children {
-		rv = append(rv, childName)
-	}
-	return rv, nil
-}
-
-`)
+	_, err := fmt.Fprint(w, tmplFuncAssetDir)
 	if err != nil {
 		return err
 	}
@@ -195,8 +155,8 @@ func getLongestAssetNameLen(toc []Asset) (longest int) {
 }
 
 // writeTOC writes the table of contents file.
-func writeTOC(w io.Writer, toc []Asset) error {
-	err := writeTOCHeader(w)
+func writeTOC(w io.Writer, toc []Asset) (err error) {
+	_, err = fmt.Fprint(w, tmplFuncAsset)
 	if err != nil {
 		return err
 	}
@@ -210,67 +170,9 @@ func writeTOC(w io.Writer, toc []Asset) error {
 		}
 	}
 
-	return writeTOCFooter(w)
-}
+	_, err = fmt.Fprint(w, "}\n")
 
-// writeTOCHeader writes the table of contents file header.
-func writeTOCHeader(w io.Writer) error {
-	_, err := fmt.Fprintf(w, `// Asset loads and returns the asset for the given name.
-// It returns an error if the asset could not be found or
-// could not be loaded.
-func Asset(name string) ([]byte, error) {
-	cannonicalName := strings.Replace(name, "\\", "/", -1)
-	if f, ok := _bindata[cannonicalName]; ok {
-		a, err := f()
-		if err != nil {
-			return nil, fmt.Errorf("Asset %%s can't read by error: %%v", name, err)
-		}
-		return a.bytes, nil
-	}
-	return nil, &os.PathError{Op: "open", Path: name, Err: os.ErrNotExist}
-}
-
-// MustAsset is like Asset but panics when Asset would return an error.
-// It simplifies safe initialization of global variables.
-// nolint: deadcode
-func MustAsset(name string) []byte {
-	a, err := Asset(name)
-	if err != nil {
-		panic("asset: Asset(" + name + "): " + err.Error())
-	}
-
-	return a
-}
-
-// AssetInfo loads and returns the asset info for the given name.
-// It returns an error if the asset could not be found or
-// could not be loaded.
-func AssetInfo(name string) (os.FileInfo, error) {
-	cannonicalName := strings.Replace(name, "\\", "/", -1)
-	if f, ok := _bindata[cannonicalName]; ok {
-		a, err := f()
-		if err != nil {
-			return nil, fmt.Errorf("AssetInfo %%s can't read by error: %%v", name, err)
-		}
-		return a.info, nil
-	}
-	return nil, &os.PathError{Op: "open", Path: name, Err: os.ErrNotExist}
-}
-
-// AssetNames returns the names of the assets.
-// nolint: deadcode
-func AssetNames() []string {
-	names := make([]string, 0, len(_bindata))
-	for name := range _bindata {
-		names = append(names, name)
-	}
-	return names
-}
-
-// _bindata is a table, holding each asset generator, mapped to its name.
-var _bindata = map[string]func() (*asset, error){
-`)
-	return err
+	return
 }
 
 // writeTOCAsset write a TOC entry for the given asset.
@@ -286,12 +188,4 @@ func writeTOCAsset(w io.Writer, asset *Asset, longestNameLen int) (err error) {
 	_, err = io.WriteString(w, toWrite)
 
 	return
-}
-
-// writeTOCFooter writes the table of contents file footer.
-func writeTOCFooter(w io.Writer) error {
-	_, err := fmt.Fprintf(w, `}
-
-`)
-	return err
 }
